Rename id to entity_id
