--- entities/database_migrations/versions/20230720_213718_initial_table_creation.py
@@ -1,40 +1,0 @@
-"""Initial table creation
-
-Create Date: 2023-07-20 21:37:18.457151
-
-"""
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '20230720_213718'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('sample',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('location', sa.String(), nullable=False),
-    sa.PrimaryKeyConstraint('id', name=op.f('pk_sample'))
-    )
-    op.create_table('sequencing_read',
-    sa.Column('sequencing_read_id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('nucleotide', sa.String(), nullable=False),
-    sa.Column('sequence', sa.String(), nullable=False),
-    sa.Column('protocol', sa.String(), nullable=False),
-    sa.Column('sample_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['sample_id'], ['sample.id'], name=op.f('fk_sequencing_read_sample_id_sample')),
-    sa.PrimaryKeyConstraint('sequencing_read_id', name=op.f('pk_sequencing_read'))
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('sequencing_read')
-    op.drop_table('sample')
-    # ### end Alembic commands ###
