--- entities/database_migrations/versions/20230731_205203_.py
@@ -1,39 +1,0 @@
-"""
-
-Create Date: 2023-07-31 20:52:04.026128
-
-"""
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '20230731_205203'
-down_revision = '20230720_213718'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('entity',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('type', sa.String(), nullable=False),
-    sa.Column('producing_run_id', sa.Integer(), nullable=True),
-    sa.Column('owner_user_id', sa.Integer(), nullable=True),
-    sa.PrimaryKeyConstraint('id', name=op.f('pk_entity'))
-    )
-    op.create_foreign_key(op.f('fk_sample_id_entity'), 'sample', 'entity', ['id'], ['id'])
-    op.add_column('sequencing_read', sa.Column('id', sa.Integer(), nullable=False))
-    op.create_foreign_key(op.f('fk_sequencing_read_id_entity'), 'sequencing_read', 'entity', ['id'], ['id'])
-    op.drop_column('sequencing_read', 'sequencing_read_id')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('sequencing_read', sa.Column('sequencing_read_id', sa.INTEGER(), autoincrement=True, nullable=False))
-    op.drop_constraint(op.f('fk_sequencing_read_id_entity'), 'sequencing_read', type_='foreignkey')
-    op.drop_column('sequencing_read', 'id')
-    op.drop_constraint(op.f('fk_sample_id_entity'), 'sample', type_='foreignkey')
-    op.drop_table('entity')
-    # ### end Alembic commands ###
